% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit the pre-defined Neural Network for Longitudinal Data}
\usage{
fit(
  model,
  ver,
  n_epoch,
  bsize,
  X1,
  X2,
  X3,
  X4,
  X5,
  X6,
  X7,
  X8,
  X9,
  X10,
  Xif,
  y
)
}
\arguments{
\item{model}{The model object produced by model().}

\item{ver}{ver=0 to show nothing, ver=1 to show animated progress bar, ver=2 to just mention the number of epoch during training.}

\item{n_epoch}{The number of epochs to train the model.}

\item{bsize}{The batch size.}

\item{X1}{Features as inputs of 1st LSTM.}

\item{X2}{Features as inputs of 2nd LSTM.}

\item{X3}{Features as inputs of 3rd LSTM.}

\item{X4}{Features as inputs of 4th LSTM.}

\item{X5}{Features as inputs of 5th LSTM.}

\item{X6}{Features as inputs of 6th LSTM.}

\item{X7}{Features as inputs of 7th LSTM.}

\item{X8}{Features as inputs of 8th LSTM.}

\item{X9}{Features as inputs of 9th LSTM.}

\item{X10}{Features as inputs of 10th LSTM.}

\item{Xif}{The features to be concatenated with the outputs of the LSTMs.}

\item{y}{The target variable.}
}
\value{
The fitted model.
}
\description{
DESCRIPTION TO BE INCLUDED
}
\examples{
fit(model = model,
    ver = 0,
    n_epochs = 200,
    bsize = 32,
    X1 = X1, 
    X2 = X2, 
    X3 = X3,
    X4 = X4,
    X5 = X5,
    X6 = X6,
    X7 = X7,
    X8 = X8,
    X9 = X9,
    X10 = X10,
    Xif = Xif,
    y = y)
}
